Render Seats for a Venue with SVG elements
For comparison [see this commit][commit]

There's nothing spectacular to show off in this commit, but perhaps thats the
point.

Breezy is really a multi-page application. We include all the content
together on the same `index.json.props` just as we would when we're
working with just a single `index.html.erb`. Then we modify the JSX markup to
account for the new incoming props.

Of note! The very next commit is a cherry-pick of the test on the original
select-your-own-seat branch. There were no additional changes to make the
tests work!

[commit]: https://github.com/seanpdoyle/select-your-own-seat/commit/216b59c57382d8d091ee380dac22779bce0f03f3