Add `seats#show` route and action
For comparison [see this commit][commit]

This changes here has a lot in common with the referenced commit. The usual
routes, controllers, models, and tests remain unchanged. There are some
differences in the views worth mentioning.

1. `SeatDialog.jsx` vs. `_dialog.html.erb` - React components vs. partials +
Stimulus. They're both using the same dialogPolyfill, ergronmically they're not
too dissimilar. The stimilus version uses the DOM api, the React version uses
its lifecycle.

A quality of the Stimulus version is it immediately starts out as HTML, and
behavior is added as sprinkles of javascript.

A quality of the React version is its testability, and its tendency to encourage
grouping of behavior into sharable, reusable components.

Its important to note that in both version, the number of HTTP request is
exactly the same. This is important because we often see companion React
libraries make another round trip for data. HTML enjoys the ability to fast
fail, either you get the entire page or you don't. With Breezy, this ability to
fast fail is brought to React.

2. There is no `show.jsx`
In Breezy applications, there is usually a `jsx` companion to the `props`
template, but in this case it was left out intentionally. Instead, we reuse
the `index.jsx` because the only difference between the `index` and the `show`
is the modal. Here's the relevant line:

```
 const identifierToComponentMapping = {
-  'seats/index': SeatsIndex
+  'seats/index': SeatsIndex,
+  'seats/show': SeatsIndex,
 }
```

[commit]: https://github.com/seanpdoyle/select-your-own-seat/commit/cd3d29c6717df5e6b169c8be211f6e2447158d74