Persist map across page loads
For comparison [see this commit][commit]

Interestingly, Breezy doesn't have this problem. Recall earlier, that we used
the same `index.jsx` component for the `show` and `index` action.

```
 const identifierToComponentMapping = {
  'seats/index': SeatsIndex,
  'seats/show': SeatsIndex,
 }
```

Because of this, we can rely on React's VDOM diffing to just change the DOM
nodes that needs to change.

We do have to reset the map when ever we receive props to achieve the same
effect when switching floors. And with that we introduce the first use of
Breezy's UJS helpers, here's the relevant diff:

```
- <RailsTag html={ activeLink }>
+ <RailsTag html={activeLink} data-bz-visit={true} />
```

Breezy's UJS helpers are similar to Rails, but you must explicitly opt-in by
adding a data attribute. Options include `data-bz-remote`, and `data-bz-visit`,
the difference being that `data-bz-visit` will push history state.

These UJS helpers are conviences to using Breezy's out-of-the-box thunks:
`visit` and `remote`. If you ever need fine grain control of the way your
application fetches, these thunks are passed as props to your page
component.

[commit]: https://github.com/seanpdoyle/select-your-own-seat/commit/bf3d350c8212cc390c068791b3831b9e465ae66f