Navigate when `<dialog>` element closes
For comparison [see this commit][commit]

This higlights the previous commit's comment on reverting the state with
`history.back`. This feature was easier to implement because Breezy stores the
previous pages state in the Redux store, and if it doesn't exist the browser
will fallback to doing a full refresh.

[commit]: https://github.com/seanpdoyle/select-your-own-seat/commit/c7613b1cd691655c014ba1d2aef935a21f38889d