"Progressively" enhance Selection State
For comparison [see this commit][commit]

Note that "Progressively" is in quotes, because while this represents a small
iterative change that uses UJS, i'm not sure if it falls in the definition of
"Progressive" enhancement.

Much like an earlier commit, `Persist map across page loads`, Breezy does not
have the same problem that the Stimulus version does.

In Breezy, we just have to make sure the application has a new state, and React
will declaratively make diffed changes to the DOM to reflect that state.

Breezy does have a different problem which is relevant to this commit message.
By default, links are just links, if we want to enhance them, we must do so
explicitly. So this commit adds `data-bz-visit`:
  1. To cart removal
  2. To the add to cart from dialog
  2. To the seat clicks so SPA transition works

[commit]: https://github.com/seanpdoyle/select-your-own-seat/commit/57347d3165acd7783c6dbf49e78e6f021ab3ff98