Dynamically Hide and Show Loading Spinner
For comparison [see this commit][commit]

This commit is a slight divergence from Stimulus. Breezy does not have global
callbacks like `turbolinks:request-start`, if you need to do something before a
request starts, you must do so explicitly.

The behavior of showing a loading indicator requires an optimistic update. We
have to change the loading state to `true` before we make a request, and
change it to `false` when we're done with the request.

To make this feature work, add your customizations to `application_visit`, which
was generated when you ran Breezy's generators then add custom action_creators,
and reducers.

Breezy leans on Redux to give you the option of creating your own action
creators, and reducers. The Breezy installer already sets up a place where you
can put this logic.

For our purposes, we need to set the loading indicator on the current page to
false. Breezy's state shape is like Turbolink's internal state, a hash where the
urls are the keys and the values are the pages's content. We also recieve the
current key from the passed in props as `this.props.pageKey`.

Once we've created our action creator, we dispatch it with the `pageKey` to
toggle the current page's loading state, here's the meat of the functionality:

```
case SHOW_LOADING: {
  const {pageKey} = action.payload

  return produce(state, draft => {
    const { seatingMap } = draft[pageKey].data
    seatingMap.loading = true
  })
}
```

Note that Breezy will add `immer` to your `package.json`.

[commit]: https://github.com/seanpdoyle/select-your-own-seat/commit/71697874c5109f55405a9d2ab29a538ed1f12c71