Dynamically Hide and Show Loading Spinner
For comparison [see this commit][commit]

This commit is a complete divergence from Stimulus. Breezy does not have global
callbacks like `turbolinks:request-start`, if you need to do something before a
request starts, you must do so explicitly.

Instead of using the UJS attribute, we have to build our own click handler by
using the `visit` thunk, which is what powers `data-bz-visit`.

Additionally, the behavior of showing a loading indicator requires an
optimistic update. We have to change the loading state to `true` before we
make a request, and change it to `false` when we're done with the request
for more advanced behavior that the UJS attributes can not handle, you can
use equivalent thunks.

Breezy leans on Redux to give you the option of creating your own action
creators, and reducers. The Breezy installer already sets up a place where you
can put this logic.

For our purposes, we need to set the loading indicator on the current page to
false. Breezy's state shape is like Turbolink's internal state, a hash where the
urls are the keys and the values are the pages's content. We also recieve the
current key from the passed in props as `this.props.pageKey`.

Once we've created our action creator, we dispatch it with the `pageKey` to
toggle the current page's loading state, here's the meat of the functionality:

```
case TOGGLE_MAP_LOADING: {
  const {pageKey} = action.payload

  return produce(state, draft => {
    const { seatingMap } = draft[pageKey].data
    seatingMap.loading = !seatingMap.loading
  })
}
```

Note that Breezy will add `immer` to your `package.json`.

[commit]: https://github.com/seanpdoyle/select-your-own-seat/commit/3a2866d7c1444bb0049f57b520e3181051cfea01