Render the modal in isolation [NEW]
When clicking a seat from the index page you're directed to the seat's `show`
page where the only real difference is the dialog modal. The current example in
Breezy, and in the Turbolinks/Stimulus version requires a full load of the
`show` page. While fine now, we can imagine a scenario where, as more visual
elements are added to both pages, a full page load would be borderline costly.

Here we begin to highlight Breezy's more unique features. Your `show.json.props`
is a queryable content tree. It is possible to pass in a path to a target node,
traverse the template tree you built, while skipping execution of nodes that do
not matter.

To do this we append a `bzq` parameter in a link enabled with `data-bzq-visit`
like so:

```
  <a href='/venues/benedum_center/floors/orchestra/seats/B-103?bzq=data.seat'
    data-bzq-visit={true}
    data-bz-placeholder={this.props.pageKey}
  />
```

or using the equivalent provided thunk like so:

```
  this.props.visit('/venues/benedum_center/floors/orchestra/seats/B-103?bzq=data.seat', {
    placeholderKey: this.props.pageKey
  })
```

In both scenarios, a placeholderKey is passed to the `visit` thunk, allowing
you to copy a specified page to use a placeholder for the next page. Then
Breezy fetches the node from the server, and finally grafts it onto the copy at
the same location.

To achieve this effect in Stimulus, you would have to create a partial that
serves the purpose of rendering with `show.html.erb` when you land on the
show page, AND rendering in isolation via a new controller, and new routes so
that it can be sent back to the client.