Add `inline_svg` gem
For comparison [see this commit][commit]

With Breezy, Using React does not mean you have to abandon all the tools that
Rails has to offer. In this commit, we introduce the `RailsTag` component that
ships with Breezy.

With this component, almost all rails view helpers can continue to add value to
your workflow. The `RailsTag` works like React's `dangerouslySetInnerHtml`,
except it does not need a wrapper div. It takes the HTML node generated by your
view helper and uses that as the wrapper div. React's click handlers will work as
expected, go ahead and add `onClick` or `onSubmit` if its a form

[commit]: https://github.com/seanpdoyle/select-your-own-seat/commit/2eb88c2b9f33cb84f4b35bfc47b484f4f7b84ca2