Opt out of preview caching
For comparison [see this commit][commit]

This is intentionally an empty commit. The problem that the reference commit
tries to solve is that Turbolink's cache, while extremely useful has limitations
that would require it to turn it off application wide.

In the referenced commit's case, the flashing of the page caused by Turbolink's
swapping of cached HTML bodys is enough to turn off this feature.

Breezy does not have this problem, because Breezy works by swapping one page's
JSON for another, then leaving it to React to render.

[commit]: https://github.com/seanpdoyle/select-your-own-seat/commit/136f68e015b18d1c2de6976cee0f456d5221b599